+++ Constraints/Count.php
@@ -72,8 +72,6 @@ class Count extends Constraint
             $exactly = $options['value'] ?? null;
         } elseif (\is_array($options)) {
             trigger_deprecation('symfony/validator', '7.3', 'Passing an array of options to configure the "%s" constraint is deprecated, use named arguments instead.', static::class);
-        } else {
-            $options = [];
         }
 
         $min ??= $options['min'] ?? null;