+++ Constraints/CssColor.php
@@ -73,7 +73,7 @@ class CssColor extends Constraint
         $validationModesAsString = implode(', ', self::$validationModes);
 
         if (!$formats) {
-            $options['value'] = self::$validationModes;
+            $formats = self::$validationModes;
         } elseif (\is_array($formats) && \is_string(key($formats))) {
             trigger_deprecation('symfony/validator', '7.3', 'Passing an array of options to configure the "%s" constraint is deprecated, use named arguments instead.', static::class);
 
@@ -82,28 +82,33 @@ class CssColor extends Constraint
             if ([] === array_intersect(self::$validationModes, $formats)) {
                 throw new InvalidArgumentException(\sprintf('The "formats" parameter value is not valid. It must contain one or more of the following values: "%s".', $validationModesAsString));
             }
-
-            $options['value'] = $formats;
         } elseif (\is_string($formats)) {
             if (!\in_array($formats, self::$validationModes, true)) {
                 throw new InvalidArgumentException(\sprintf('The "formats" parameter value is not valid. It must contain one or more of the following values: "%s".', $validationModesAsString));
             }
 
-            $options['value'] = [$formats];
+            $formats = [$formats];
         } else {
             throw new InvalidArgumentException('The "formats" parameter type is not valid. It should be a string or an array.');
         }
 
         parent::__construct($options, $groups, $payload);
 
+        $this->formats = $formats ?? $this->formats;
         $this->message = $message ?? $this->message;
     }
 
+    /**
+     * @deprecated since Symfony 7.4
+     */
     public function getDefaultOption(): string
     {
         return 'formats';
     }
 
+    /**
+     * @deprecated since Symfony 7.4
+     */
     public function getRequiredOptions(): array
     {
         return ['formats'];