+++ Constraints/NotBlank.php
@@ -47,7 +47,7 @@ class NotBlank extends Constraint
             trigger_deprecation('symfony/validator', '7.3', 'Passing an array of options to configure the "%s" constraint is deprecated, use named arguments instead.', static::class);
         }
 
-        parent::__construct($options ?? [], $groups, $payload);
+        parent::__construct($options, $groups, $payload);
 
         $this->message = $message ?? $this->message;
         $this->allowNull = $allowNull ?? $this->allowNull;